Follow up issue #140 

Add interference edge pruning esterel plan / graph optimization based on Floyd-Warshall algorithm.
There are two levels of optimization: normal and intense
The difference is whether you update the distance matrix every time you add
an interference edge or one time at the beginning only.

Here is a benchmark on problem 

P15-15-3-1-1 from mapanalyzer domain:

problem | time [s] with all floyd-warshall | time [s] with one time floyd-warshall | time [s] without floyd-warshall | with getBounds
-- | -- | -- | -- | --
P15-15-3-1-1 | 47 | 35 | 35 | 58

NOTE: with all floyd-warshall means every time we add an interference edge, we update the distance matrix (which is an expensive operation).